refactor: . Move createEditingInstructionForGroup() to own file, nearer code it uses
